feat: alternative layout approach using flexbox (#38)
BREAKING CHANGE: We are now using Flexbox because it works better with dynamically hiding and showing fields using the `show` schema property. This is a breaking change because forms using the previous Grid layout system would leave an empty place in the grid when they were hidden, but the new system will stretch rows to fill in missing space.